pyproject.toml
[tool.poetry]
name = "marketplace"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
fastapi = "^0.109.2"
python-multipart = "^0.0.7"
pydantic = ">2.0"
gunicorn = "^21.2.0"
alembic = "^1.12.1"
psycopg = {extras = ["binary"], version = "^3.1.13"}
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mypy = "^1.8.0"
ruff = "^0.2.2"
pre-commit = "^3.6.2"
coverage = "^7.4.3"
httpx = "^0.27.0"
pytest-env = "^1.1.3"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.6"
factory-boy = "^3.3.0"
async-factory-boy = "^1.0.1"
asgi-lifespan = "^2.1.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.mypy]
strict = true
exclude = ["venv", "alembic"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true